+++ docker/manylinux/build_wheels.sh
@@ -5,7 +5,7 @@ set -e -x
 GRANNE_DIR="/granne/"
 
 export LD_LIBRARY_PATH=/usr/local/lib/:$LD_LIBRARY_PATH
-for PYBIN in /opt/python/cp{37,36}*/bin; do
+for PYBIN in /opt/python/cp{37,36,35}*/bin; do
     (
         export PATH="$PYBIN:$PATH"
         pip install setuptools setuptools-rust