+++ CHANGELOG.md
@@ -7,4 +7,4 @@ Major rewrite. Changes include:
 * Simplified API
 * Documentation
 * Improved python bindings
-* File formats for index and elements are incompatible with previous versions
+* Note: File formats for index and elements are incompatible with previous versions